--- src/main/java/com/wpaul15/scryfall/api/query/IExpression.java
@@ -1,3 +1,0 @@
-package com.wpaul15.scryfall.api.query;
-
-public interface IExpression<T> extends IQueryParams {}
