+++ src/main/java/com/wpaul15/scryfall/api/query/IFilter.java
@@ -1,0 +1,3 @@
+package com.wpaul15.scryfall.api.query;
+
+public interface IFilter<T> extends IQueryParams {}
