+++ src/main/java/com/wpaul15/scryfall/api/query/filter/AbstractFilter.java
@@ -1,13 +1,13 @@
-package com.wpaul15.scryfall.api.query.expression;
+package com.wpaul15.scryfall.api.query.filter;
 
-import com.wpaul15.scryfall.api.query.IExpression;
+import com.wpaul15.scryfall.api.query.IFilter;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PROTECTED)
-abstract class AbstractExpression<T> implements IExpression<T> {
+abstract class AbstractFilter<T> implements IFilter<T> {
 
   Iterable<T> entries;
